Console-based UI
Implemented a rudimentary console-based user interface for the purpose
of testing the game logic. This simplified version of the game uses
letters to represent chess pieces (and periods for empty tiles),
prompts users for correct tile inputs, and displays the current board
and turn. Also prompts user to make valid moves.