Reconfigured chess piece objects
The objects for each chess piece now represent an individual piece on
the board, thus they remember their row number, column number, and
color. Initially, new piece objects were created for every time a piece
was moved whereas now, all the pieces are only created once: at the
beginning of the game. This allows the game state and the pieces to
maintain a history of their activity.