Castling
Both players can castle king-side and queen-side. Castling can only
occur if the king and rook being flipped have not moved during the
game. Also, their original tiles and tiles they move across must not be
in the line of vision for any of the opponent pieces. Castling is
allowed to result in checkmate.