Tile Highlighting
Tile is highlighted when user selects own piece to move. You can switch
the piece designated to move by clicking on another one of your pieces.
Also fixed an error that would cause the GUI to crash after an invalid
move. When the user makes an invalid move, nothing happens.